refactor: move publishing workflow tests
